+++ lib/xwidget.dart
@@ -3,6 +3,8 @@ export 'src/el/parser.dart';
 export 'src/custom/async.dart';
 export 'src/custom/collection.dart';
 export 'src/custom/controller.dart';
+export 'src/custom/event_listener.dart';
+export 'src/custom/material_state.dart';
 export 'src/custom/media.dart';
 export 'src/custom/value_listener.dart';